requirements.txt
gymnasium
stable-baselines3[extra]
mss
pillow
pynput
opencv-python
torch
numpy